+++ mozrepl/type/__init__.py
@@ -6,3 +6,4 @@ from __future__ import absolute_import
 from .object import Object
 from .function import Function
 from .array import Array
+from .raw import Raw