+++ mozrepl/util.py
@@ -11,10 +11,8 @@ def convertToJs(arg):
 	"""
 	입력받은 값을 javascript에서 사용 가능한 값으로 변환합니다.
 	
-	.. todo:: dict 형태 변환 기능을 추가.
-	
 	:param arg: 변환할 값. 만약 list와 tuple타입을 인자로 준다면, 포함된 값은 :py:func:`~mozrepl.util.convertToJs` 함수가 변환 할 수 있는 값이어야 합니다.
-	:type arg: int, float, None, unicode, bytes, str, tuple, list, :py:class:`~mozrepl.type.Object`, :py:class:`~mozrepl.type.Function`, :py:class:`~mozrepl.type.Array`
+	:type arg: int, float, None, unicode, bytes, str, dict, tuple, list, :py:class:`~mozrepl.type.Object`, :py:class:`~mozrepl.type.Function`, :py:class:`~mozrepl.type.Array`, :py:class:`~mozrepl.type.Raw`
 	:return: javascript에서 사용 할 수 있는 값.
 	:rtype: unicode
 	"""
@@ -30,13 +28,22 @@ def convertToJs(arg):
 	if isinstance(arg, str):
 		return repr(arg)
 	
+	if isinstance(arg, dict):
+		buffer = lambda (k, v): '{k}: {v}'.format(k=convertToJs(k), v=convertToJs(v))
+		buffer = map(buffer, arg.items())
+		buffer = ', '.join(buffer)
+		return '{{{0}}}'.format(buffer)
+	
 	if isinstance(arg, (tuple, list)):
 		buffer = map(convertToJs, arg)
 		buffer = ', '.join(buffer)
 		return '[{0}]'.format(buffer)
 	
-	from .type import Object
+	from .type import Object, Raw
 	if isinstance(arg, Object):
+		return unicode(arg)
+	
+	if isinstance(arg, Raw):
 		return unicode(arg)
 	
 	raise TypeError('"{type}" 타입은 변환 할 수 없는 타입입니다.'.format(type=type(arg)))